everything else except final comments

+++ Auth_and_AC.py
@@ -2,25 +2,28 @@ from CryptoProject import CryptoProject
 from cryptography.fernet import Fernet
 import sqlite3
 from CryptoProject import CryptoProject as crypto
+import os
+
+# Password store for current user which is deleted at end of user session
+global current_pass
+current_pass = None
 
 # Initialize CryptoProject class
 crypto = CryptoProject()
 
-# File to store user accounts
-# You can implement the backing store using a database or other methods as you like
-
-
-
 class Authentication():
     def __init__(self):
+        # Initialize database
         conn = sqlite3.connect('users.db')
         cursor = conn.cursor()
 
+        # Username and password table
         cursor.execute('CREATE TABLE IF NOT EXISTS users(\
-                       username TEXT NOT NULL UNIQUE, \
+                       username TEXT NOT NULL, \
                        password TEXT NOT NULL)')
         
-        cursor.execute('CREATE TABLE IF NOT EXISTS acl(doc_name TEXT NOT NULL UNIQUE, username TEXT NOT NULL)')
+        # ACL table
+        cursor.execute('CREATE TABLE IF NOT EXISTS acl(doc_name TEXT NOT NULL, username TEXT NOT NULL)')
 
         conn.commit()
         
@@ -29,59 +32,69 @@ class Authentication():
     
     def load_users(self):
         """Load users from persistent storage."""
-        try:
-            conn = sqlite3.connect('users.db')
+        conn = sqlite3.connect('users.db')
 
-            cursor = conn.cursor()
-            cursor.execute('SELECT * FROM users')
+        cursor = conn.cursor()
+        cursor.execute('SELECT * FROM users')
 
-            users = cursor.fetchall()
+        users = cursor.fetchall()
 
-            conn.close()
+        conn.close()
 
-            return users
-        except FileNotFoundError:
-            return {}
+        return users
 
     def save_users(self, users):
-        # Feed in a dictionary with keys of username and values of password
+        # Insert new user into database
         conn = sqlite3.connect('users.db')
         cursor = conn.cursor()
-        cursor.executemany("INSERT INTO users VALUES(:username, :password)", users)
+        cursor.executemany("INSERT INTO users VALUES(?, ?)", (users,))
 
         conn.commit()
         conn.close()
         return
 
     def create_account(self, users):
-        
-        # TODO: Implement account creation
         username = input("Create a username: ")
         password = input("Create a password: ")
-        # TODO: Check if username already exists
+        
+        # Check if username already exists
         for user in users:
             if user[0] == username:
                 print("Username already exists.")
                 return
-        # TODO: Store password securely
+
+        # Hash password
         hash = crypto.hash_string(password)
-
+        
+        # Generate RSA keys
         crypto.generate_rsa_keys(username)
+        
+        # Store RSA private key encrypted with AES
+        with open(username + '_private.pem', 'r') as f:
+            private_key = f.read()
 
-        # TODO: Save updated user list
-        data = ({'username': username, 'password': hash},)
+        hidden_key = crypto.aes_encrypt(private_key, password)
+
+        with open(username + '_private.pem', 'w') as f:
+            f.write(hidden_key)
+
+        # Save user to persistent storage
+        data = (username, hash)
         self.save_users(data)
         return
 
     def login(self, users):
+        global current_pass
         username = input("Enter username: ")
         password = input("Enter password: ")
-        # TODO: Implement login method including secure password check
+        
+        # Check if username and password match
         for user in users:
             if user[0] == username:
                 hash = user[1]
                 if crypto.verify_integrity(password, hash):
                     print("Login successful.")
+                    current_pass = password
                     return username
         print("Invalid login credentials")
         return False
@@ -92,6 +105,7 @@ class AccessControl():
         return
 
     def load_acl(self):
+        # Load ACL from persistent storage
         conn = sqlite3.connect('users.db')
 
         cursor = conn.cursor()
@@ -103,38 +117,34 @@ class AccessControl():
         
         if acl:
             return acl
-        # TODO: Load ACL (Access Control List) from persistent storage.
         return []
 
     def save_acl(self, acl):
+        # Insert new document with associated user into database
         conn = sqlite3.connect('users.db')
         cursor = conn.cursor()
         cursor.executemany("INSERT INTO acl VALUES(?, ?)", (acl, ))
 
         conn.commit()
         conn.close()
-        #TODO: Save ACL to persistent storage.
         return
 
-    def create_file(self, username):
+    def create_file(self, username, acl):
         filename = input("Enter the name of the file you want to create: ")
         content = input("Enter content for the file: ")
+        
+        path = os.getcwd()
 
-        encrypted_content = crypto.rsa_encrypt(content, username)
-        # TODO: Create the file and write content. EXTRA CREDIT: encrypt the file/content.
-        with open(filename + '.txt', 'w') as f:
-            f.write(encrypted_content)
-        # TODO: Add file access entry in ACL
+        # Create file with encypted content if it doesn't exist
+        if not os.path.exists(path + "/" + filename + '.txt'):
+            encrypted_content = crypto.rsa_encrypt(content, username)
+            with open(filename + '.txt', 'w') as f:
+                f.write(encrypted_content)
+            new_acl = (filename, username)
+            self.save_acl(new_acl)
+            return
 
-        acl = (filename, username)
-        self.save_acl(acl)
-        return
-
-
-
-    def read_file(self, username, acl):
-        filename = input("Enter the name of the file you want to read: ")
-
+        # If file already exists, check if user has access
         for file in acl:
             if file[0] == filename:
                 if file[1] == username:
@@ -143,15 +153,43 @@ class AccessControl():
                     print("You do not have access to this file.")
                     return
 
-        decrypted_content = crypto.rsa_decrypt(filename, username)
-        print(decrypted_content)
-        
-        # TODO: Check if the user has access. EXTRA CREDIT: If file was encrypted, decrypt the file/content
+        # Ask if user wants to overwrite existing file that they own
+        while True:
+            overwrite = input("File already exists. Do you want to overwrite it? (y/n) ")
+            if overwrite.lower() == 'y':
+                encrypted_content = crypto.rsa_encrypt(content, username)
+                with open(filename + '.txt', 'w') as f:
+                    f.write(encrypted_content)
+                new_acl = (filename, username)
+                self.save_acl(new_acl)
+                return
+            elif overwrite.lower() == 'n':
+                return
+            else:
+                print("Invalid input.")
 
-        # TODO: Optionally decrypt the file content
+    def read_file(self, username, acl):
+        filename = input("Enter the name of the file you want to read: ")
 
+        # Check if user has access        
+        for file in acl:
+            if file[0] == filename:
+                if file[1] == username:
+                    break
+                else:
+                    print("You do not have access to this file.")
+                    return
+
+        # Read decrypted file if user has access
+        try:
+            decrypted_content = crypto.rsa_decrypt(filename, username, current_pass)
+            print(decrypted_content)
+        except FileNotFoundError:
+            print("File does not exist.")
+            return
 
 def main():
+    global current_pass
     auth = Authentication()
     ac = AccessControl()
     
@@ -181,18 +219,38 @@ def main():
                     file_choice = input("Enter your choice: ")
                     
                     if file_choice == '1':
-                        ac.create_file(user)
+                        ac.create_file(user, acl)
+                        acl = ac.load_acl()
                     elif file_choice == '2':
                         ac.read_file(user, acl)
                     elif file_choice == '3':
                         print(f"Logging out {user}.")
+                        current_pass = None
                         break
                     else:
                         print("Invalid choice.")
         elif choice == '3':
+            current_pass = None
             break
         else:
             print("Invalid choice.")
 
 if __name__ == "__main__":
     main()
+
+
+"""
+DATABASE SETUP: SQLite with two tables (users and acl)
+PASSWORD STORE: Passwords are hashed and stored in the database
+    This follows the Principle of Least Privilege because users do not have access to their own password
+    Hashing passwords mitigates risk in case of a data breach
+USER AND AC LISTS: Users and Access Control lists are stored in the database
+    This follows the Principle of Least Privilege because only the program can edit these lists
+    Using the database prevents attackers from editing the lists via creation of accounts or logins in the program
+RSA KEYS: Keys are stored in files, and private keys are encrypted with AES using the user password
+    This follows the Principle of Open Design because everything, including the files, are visible to any user, but 
+        none are readable exepct through the proper user login
+    This prevents an attacker from gaining access to the files and directly using the content because all important
+        information is encrypted, including RSA keys using AES
+
+"""
